move timer functions to timer.go and tests to timer_test.go
